Finishing removal of deprecated pivoting/permutation routines (with the exception of PivotsToPartialPermutation, which is not currently used by Permutation and DistPermutation)
