+++ include/El/lapack_like/perm.hpp
@@ -15,218 +15,6 @@
 
 namespace El {
 
-// Deprecation line
-// ################
-
-// Permute columns
-// ===============
-// TODO: Deprecate this
-template<typename T>
-void PermuteCols
-(       AbstractDistMatrix<T>& A,
-  const PermutationMeta& meta,
-  bool inverse=false ); 
-
-// Permute rows
-// ============
-// TODO: Deprecate this
-template<typename T>
-void PermuteRows
-(       AbstractDistMatrix<T>& A,
-  const PermutationMeta& meta,
-  bool inverse=false ); 
-
-// Symmetric permutation
-// =====================
-template<typename T>
-void SymmetricPermutation
-( UpperOrLower uplo,
-        AbstractDistMatrix<T>& A,
-  const DistPermutation& perm,
-  bool inverse=false,
-  bool conjugate=false );
-
-// Apply column pivots
-// ===================
-// NOTE: These routine are now deprecated
-
-template<typename T>
-void ApplyColPivots
-(       Matrix<T>& A,
-  const Matrix<Int>& pivots,
-        Int offset=0 );
-template<typename T>
-void ApplyColPivots
-(       AbstractDistMatrix<T>& A,
-  const AbstractDistMatrix<Int>& pivots, 
-        Int offset=0 );
-
-template<typename T>
-void ApplyInverseColPivots
-(       Matrix<T>& A,
-  const Matrix<Int>& pivots,
-        Int offset=0 );
-template<typename T>
-void ApplyInverseColPivots
-(       AbstractDistMatrix<T>& A,
-  const AbstractDistMatrix<Int>& pivots, 
-        Int offset=0 );
-
-// Apply row pivots
-// ================
-// NOTE: These routine are now deprecated
-
-template<typename T>
-void ApplyRowPivots
-(       Matrix<T>& A,
-  const Matrix<Int>& pivots,
-        Int offset=0 );
-template<typename T>
-void ApplyRowPivots
-(       AbstractDistMatrix<T>& A,
-  const AbstractDistMatrix<Int>& pivots,
-        Int offset=0 );
-
-template<typename T>
-void ApplyInverseRowPivots
-(       Matrix<T>& A,
-  const Matrix<Int>& pivots,
-        Int offset=0 );
-template<typename T>
-void ApplyInverseRowPivots
-(       AbstractDistMatrix<T>& A,
-  const AbstractDistMatrix<Int>& pivots, 
-        Int offset=0 );
-
-// Apply symmetric pivots
-// ======================
-// NOTE: These routine are now deprecated
-template<typename T>
-void ApplySymmetricPivots
-( UpperOrLower uplo,
-        Matrix<T>& A, 
-  const Matrix<Int>& p,
-  bool conjugate=false, 
-  Int offset=0 );
-template<typename T>
-void ApplySymmetricPivots
-( UpperOrLower uplo,
-        AbstractDistMatrix<T>& A,
-  const AbstractDistMatrix<Int>& pivots,
-  bool conjugate=false, 
-  Int offset=0 );
-
-template<typename T>
-void ApplyInverseSymmetricPivots
-( UpperOrLower uplo,
-        Matrix<T>& A, 
-  const Matrix<Int>& p,
-  bool conjugate=false, 
-  Int offset=0 );
-template<typename T>
-void ApplyInverseSymmetricPivots
-( UpperOrLower uplo,
-        AbstractDistMatrix<T>& A,
-  const AbstractDistMatrix<Int>& pivots,
-  bool conjugate=false, 
-  Int offset=0 );
-
-// Explicit permutation
-// ====================
-// NOTE: These routine are now deprecated
-void ExplicitPermutation
-( const Matrix<Int>& p,
-        Matrix<Int>& P );
-void ExplicitPermutation
-( const AbstractDistMatrix<Int>& p,
-        AbstractDistMatrix<Int>& P );
-
-// Invert permutation
-// ==================
-// NOTE: These routine are now deprecated
-void InvertPermutation
-( const Matrix<Int>& p,
-        Matrix<Int>& pInv );
-void InvertPermutation
-( const AbstractDistMatrix<Int>& p,
-        AbstractDistMatrix<Int>& pInv );
-
-// Parity of a permutation
-// =======================
-// NOTE: These routine are now deprecated
-bool PermutationParity( const Matrix<Int>& p );
-bool PermutationParity( const AbstractDistMatrix<Int>& p );
-
-// Permute columns
-// ===============
-// NOTE: These routine are now deprecated
-template<typename T>
-void PermuteCols
-(       Matrix<T>& A,
-  const Matrix<Int>& p );
-template<typename T>
-void PermuteCols
-(       AbstractDistMatrix<T>& A,
-  const AbstractDistMatrix<Int>& p );
-
-template<typename T>
-void InversePermuteCols
-(       Matrix<T>& A,
-  const Matrix<Int>& p );
-template<typename T>
-void InversePermuteCols
-(       AbstractDistMatrix<T>& A,
-  const AbstractDistMatrix<Int>& p );
-
-template<typename T>
-void PermuteCols
-(       Matrix<T>& A,
-  const Matrix<Int>& p,
-  const Matrix<Int>& pInv );
-template<typename T>
-void PermuteCols
-(       AbstractDistMatrix<T>& A,
-  const AbstractDistMatrix<Int>& p,
-  const AbstractDistMatrix<Int>& pInv );
-
-// Permute rows
-// ============
-// NOTE: These routine are now deprecated
-template<typename T>
-void PermuteRows
-(       Matrix<T>& A,
-  const Matrix<Int>& p );
-template<typename T>
-void PermuteRows
-(       AbstractDistMatrix<T>& A,
-  const AbstractDistMatrix<Int>& p );
-
-template<typename T>
-void InversePermuteRows
-(       Matrix<T>& A, 
-  const Matrix<Int>& p );
-template<typename T>
-void InversePermuteRows
-(       AbstractDistMatrix<T>& A,
-  const AbstractDistMatrix<Int>& p );
-
-template<typename T>
-void PermuteRows
-(       Matrix<T>& A,
-  const Matrix<Int>& p,
-  const Matrix<Int>& pInv );
-template<typename T>
-void PermuteRows
-(       AbstractDistMatrix<T>& A,
-  const AbstractDistMatrix<Int>& p,
-  const AbstractDistMatrix<Int>& pInv );
-
-// Parity of a sequence of pivots
-// ==============================
-// NOTE: These routine are now deprecated
-bool PivotParity( const Matrix<Int>& p, Int pivotOffset=0 );
-bool PivotParity( const AbstractDistMatrix<Int>& p, Int pivotOffset=0 );
-
 // Convert a pivot sequence to a partial permutation vector
 // ========================================================
 // NOTE: These routine are now deprecated
@@ -243,27 +31,6 @@ void PivotsToPartialPermutation
 void PivotsToPartialPermutation
 ( const DistMatrix<Int,STAR,STAR>& pivots,
         AbstractDistMatrix<Int>& p,
-        AbstractDistMatrix<Int>& pInv,
-        Int offset=0 );
-
-// Convert a pivot sequence to a permutation vector
-// ================================================
-// NOTE: These routine are now deprecated
-void PivotsToPermutation
-( const Matrix<Int>& pivots,
-        Matrix<Int>& p,
-        Int offset=0 );
-void PivotsToPermutation
-( const AbstractDistMatrix<Int>& pivots,
-        AbstractDistMatrix<Int>& p,
-        Int offset=0 );
-
-void PivotsToInversePermutation
-( const Matrix<Int>& pivots,
-        Matrix<Int>& pInv,
-        Int offset=0 );
-void PivotsToInversePermutation
-( const AbstractDistMatrix<Int>& pivots,
         AbstractDistMatrix<Int>& pInv,
         Int offset=0 );
