+++ src/lapack_like/perm/Permutation.cpp
@@ -10,6 +10,152 @@
 
 namespace El {
 
+namespace {
+
+template<typename T>
+inline void PermuteCols
+(       Matrix<T>& A,
+  const Matrix<Int>& perm,
+  const Matrix<Int>& invPerm )
+{
+    const Int b = perm.Height();
+    DEBUG_ONLY(
+      CSE cse("PermuteCols");
+      if( A.Width() < b || b != invPerm.Height() )
+          LogicError
+          ("perm and invPerm must be vectors of equal length that are not "
+           "wider than A.");
+    )
+    const Int m = A.Height();
+    const Int n = A.Width();
+    if( m == 0 || n == 0 )
+        return;
+
+    // Make a copy of the first b columns
+    auto AColPanView = A( IR(0,m), IR(0,b) );
+    auto AColPanCopy( AColPanView );
+
+    // Make a copy of the preimage columns
+    Matrix<T> APreimageCopy( m, b );
+
+    const Int* permBuf = perm.LockedBuffer();
+    const Int* invPermBuf = invPerm.LockedBuffer();
+          T* ABuf = A.Buffer();
+          T* APreBuf = APreimageCopy.Buffer();
+          T* AColPanBuf = AColPanCopy.Buffer();
+    const Int ALDim = A.LDim();
+    const Int APreLDim = APreimageCopy.LDim();
+    const Int AColPanLDim = AColPanCopy.LDim();
+
+    for( Int j=0; j<b; ++j )
+    {
+        const Int jPre = permBuf[j];
+        if( jPre >= b )
+            MemCopy( &APreBuf[j*APreLDim], &ABuf[jPre*ALDim], m );
+    }
+
+    // Apply the permutations
+    for( Int j=0; j<b; ++j )
+    {
+        const Int jPre = permBuf[j];
+        const Int jPost = invPermBuf[j];
+        // Move row[j] into row[jPost]
+        MemCopy( &ABuf[jPost*ALDim], &AColPanBuf[j*AColPanLDim], m );
+        // Move row[jPre] into row[j]
+        if( jPre >= b )
+            MemCopy( &ABuf[j*ALDim], &APreBuf[j*APreLDim], m );
+    }
+}
+
+template<typename T>
+inline void PermuteRows
+(       Matrix<T>& A,
+  const Matrix<Int>& perm,
+  const Matrix<Int>& invPerm )
+{
+    const Int b = perm.Height();
+    DEBUG_ONLY(
+      CSE cse("PermuteRows");
+      if( A.Height() < b || b != invPerm.Height() )
+          LogicError
+          ("perm and invPerm must be vectors of equal length that are not "
+           "taller than A.");
+    )
+    const Int m = A.Height();
+    const Int n = A.Width();
+    if( m == 0 || n == 0 )
+        return;
+
+    const Int* permBuf = perm.LockedBuffer();
+    const Int* invPermBuf = invPerm.LockedBuffer();
+
+    // Make a copy of the first b rows
+    auto ARowPanView = A( IR(0,b), IR(0,n) );
+    Matrix<T> ARowPanTrans;
+    Transpose( ARowPanView, ARowPanTrans );
+
+    // Make a copy of the preimage rows
+    Matrix<T> APreimageTrans( n, b );
+
+    T* ABuf = A.Buffer();
+    T* APreBuf = APreimageTrans.Buffer();
+    T* ARowPanBuf = ARowPanTrans.Buffer();
+    const Int ALDim = A.LDim();
+    const Int APreLDim = APreimageTrans.LDim();
+    const Int ARowPanLDim = ARowPanTrans.LDim();
+
+    for( Int i=0; i<b; ++i )
+    {
+        const Int iPre = permBuf[i];
+        if( iPre >= b )
+            for( Int j=0; j<n; ++j )
+                APreBuf[j+i*APreLDim] = ABuf[iPre+j*ALDim];
+    }
+
+    // Apply the permutations
+    for( Int i=0; i<b; ++i )
+    {
+        const Int iPre = permBuf[i];
+        const Int iPost = invPermBuf[i];
+        // Move row[i] into row[image[i]]
+        for( Int j=0; j<n; ++j )
+            ABuf[iPost+j*ALDim] = ARowPanBuf[j+i*ARowPanLDim];
+        if( iPre >= b )
+        {
+            // Move row[preimage[i]] into row[i]
+            for( Int j=0; j<n; ++j )
+                ABuf[i+j*ALDim] = APreBuf[j+i*APreLDim];
+        }
+    }
+}
+
+inline void InvertPermutation
+( const Matrix<Int>& p, Matrix<Int>& pInv )
+{
+    DEBUG_ONLY(
+      CSE cse("InvertPermutation");
+      if( p.Width() != 1 )
+          LogicError("p must be a column vector");
+    )
+    const Int n = p.Height();
+    pInv.Resize( n, 1 );
+    if( n == 0 )
+        return;
+
+    DEBUG_ONLY(
+      // This is obviously necessary but not sufficient for 'p' to contain
+      // a reordering of (0,1,...,n-1).
+      const Int range = MaxNorm( p ) + 1;
+      if( range != n )
+          LogicError("Invalid permutation range");
+    )
+
+    for( Int i=0; i<n; ++i )
+        pInv.Set( p.Get(i,0), 0, i );
+}
+
+} // anonymous namespace
+
 Permutation::Permutation() { }
 
 void Permutation::Empty()
@@ -213,9 +359,41 @@ bool Permutation::Parity() const
     DEBUG_ONLY(CSE cse("Permutation::Parity"))
     if( staleParity_ )
     {
-        LogicError("Computation of parity not yet enabled");
+        if( swapSequence_ )
+            LogicError("Unexpected stale parity for a swap sequence");
+
+        // Walking through the process of LU factorization with partial 
+        // pivoting for a permutation matrix, which never requires a
+        // Schur-complement update, yields an algorithm for expressing the 
+        // inverse of a permutation in terms of a sequence of transpositions in         // linear time. Note that performing the swaps requires access to the 
+        // inverse permutation, which can be formed in linear time.
+        if( staleInverse_ )
+        {
+            El::InvertPermutation( perm_, invPerm_ );
+            staleInverse_ = false;
+        }
+
+        auto permCopy( perm_ );
+        auto invPermCopy( invPerm_ );
+
+        parity_ = false;
+        for( Int k=0; k<size_; ++k )
+        {
+            const Int permVal = permCopy.Get(k,0);
+            if( permVal != k )
+            {
+                parity_ = !parity_;
+                const Int invPermVal = invPermCopy.Get(k,0);
+                // We only need to perform half of the swaps
+                //      perm[k] <-> perm[invPerm[k]]
+                //   invPerm[k] <-> invPerm[perk[k]] 
+                // since we will not need to access perm[k] and invPerm[k] again
+                permCopy.Set( invPermVal, 0, permVal );
+                invPermCopy.Set( permVal, 0, invPermVal );
+            }
+        }
+
         staleParity_ = false;
-        return parity_; 
     }
     return parity_;
 }
@@ -270,7 +448,7 @@ void Permutation::PermuteCols( Matrix<T>& A, Int offset ) const
         // TODO: Move El::InversePermutation into this class
         if( staleInverse_ )
         {
-            InvertPermutation( perm_, invPerm_ );
+            El::InvertPermutation( perm_, invPerm_ );
             staleInverse_ = false;
         }
         // TODO: Move El::PermuteCols into this class
@@ -316,7 +494,7 @@ void Permutation::InversePermuteCols( Matrix<T>& A, Int offset ) const
         // TODO: Move El::InversePermutation into this class
         if( staleInverse_ )
         {
-            InvertPermutation( perm_, invPerm_ );
+            El::InvertPermutation( perm_, invPerm_ );
             staleInverse_ = false;
         }
         // TODO: Move El::PermuteCols into this class
@@ -359,7 +537,7 @@ void Permutation::PermuteRows( Matrix<T>& A, Int offset ) const
         // TODO: Move El::InversePermutation into this class
         if( staleInverse_ )
         {
-            InvertPermutation( perm_, invPerm_ );
+            El::InvertPermutation( perm_, invPerm_ );
             staleInverse_ = false;
         }
         // TODO: Move El::PermuteRows into this class
@@ -405,7 +583,7 @@ void Permutation::InversePermuteRows( Matrix<T>& A, Int offset ) const
         // TODO: Move El::InversePermutation into this class
         if( staleInverse_ )
         {
-            InvertPermutation( perm_, invPerm_ );
+            El::InvertPermutation( perm_, invPerm_ );
             staleInverse_ = false;
         }
         // TODO: Move El::PermuteRows into this class
@@ -455,7 +633,7 @@ void Permutation::PermuteSymmetrically
         // TODO: Move El::InversePermutation into this class
         if( staleInverse_ )
         {
-            InvertPermutation( perm_, invPerm_ );
+            El::InvertPermutation( perm_, invPerm_ );
             staleInverse_ = false;
         }
         LogicError("General symmetric permutations are not yet supported");
@@ -504,7 +682,7 @@ void Permutation::InversePermuteSymmetrically
         // TODO: Move El::InversePermutation into this class
         if( staleInverse_ )
         {
-            InvertPermutation( perm_, invPerm_ );
+            El::InvertPermutation( perm_, invPerm_ );
             staleInverse_ = false;
         }
         LogicError("General symmetric permutations are not yet supported");
@@ -514,9 +692,19 @@ void Permutation::InversePermuteSymmetrically
 void Permutation::Explicit( Matrix<Int>& P ) const
 {
     DEBUG_ONLY(CSE cse("Permutation::Explicit"))
-    // TODO: Faster algorithm
-    Identity( P, size_, size_ );
-    PermuteRows( P );
+    if( swapSequence_ )
+    {   
+        // Compose the swaps into an explicit permutation vector
+        // -----------------------------------------------------
+        perm_.Resize( size_, 1 );
+        for( Int i=0; i<size_; ++i )
+            perm_.Set( i, 0, i );
+        PermuteRows( perm_ );
+    }
+
+    Zeros( P, size_, size_ );
+    for( Int i=0; i<size_; ++i )
+        P.Set( i, perm_.Get(i,0), 1 );
 }
 
 #define PROTO(T) \
